+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/handler/driver/FixedDelayDriverProvider.java
@@ -10,7 +10,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
-import org.springframework.scheduling.support.CronTrigger;
 import org.springframework.stereotype.Component;
 
 import java.util.HashSet;
@@ -21,18 +20,18 @@ import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 
 /**
- * Cron驱动提供器。
+ * 固定间隔驱动提供器。
  *
  * @author DwArFeng
  * @since beta-1.0.0
  */
 @Component
-public class CronDriverProvider implements DriverProvider {
+public class FixedDelayDriverProvider implements DriverProvider {
 
-    public static final String SUPPORT_TYPE = "cron_driver";
+    public static final String SUPPORT_TYPE = "fixed_delay_driver";
 
     @Autowired
-    private CronDriver cronDriver;
+    private FixedDelayDriver fixedDelayDriver;
 
     @Override
     public boolean supportType(String type) {
@@ -41,7 +40,7 @@ public class CronDriverProvider implements DriverProvider {
 
     @Override
     public Driver provide() {
-        return cronDriver;
+        return fixedDelayDriver;
     }
 
     @Override
@@ -51,21 +50,21 @@ public class CronDriverProvider implements DriverProvider {
 
     @Override
     public String provideLabel() {
-        return "Cron驱动器";
+        return "固定间隔驱动器";
     }
 
     @Override
     public String provideDescription() {
-        return "根据指定的Cron表达式定时驱动的驱动器";
+        return "根据指定的间隔定时驱动，如果某一次驱动晚于间隔，则后续驱动的时间相应的顺延。";
     }
 
     @Override
     public String provideExampleContent() {
-        return "0/2 * * * * *";
+        return "60000";
     }
 
     @Component
-    public static class CronDriver implements Driver {
+    public static class FixedDelayDriver implements Driver {
 
         @Autowired
         private ThreadPoolTaskScheduler scheduler;
@@ -74,21 +73,21 @@ public class CronDriverProvider implements DriverProvider {
 
         private final Lock lock = new ReentrantLock();
         private final Set<ScheduledFuture<?>> scheduledFutures = new HashSet<>();
-        private final Set<CronProcessor> cronProcessors = new HashSet<>();
+        private final Set<FixedDelayProcessor> fixedDelayProcessors = new HashSet<>();
 
         @Override
         public void register(DriverInfo driverInfo) throws DriverException {
             lock.lock();
             try {
                 LongIdKey sectionKey = driverInfo.getSectionKey();
-                String cron = driverInfo.getContent();
-                CronProcessor cronProcessor = new CronProcessor(
+                long delay = Long.parseLong(driverInfo.getContent());
+                FixedDelayProcessor fixedDelayProcessor = new FixedDelayProcessor(
                         evaluateService,
                         sectionKey
                 );
-                CronTrigger cronTrigger = new CronTrigger(cron);
-                ScheduledFuture<?> scheduledFuture = scheduler.schedule(cronProcessor, cronTrigger);
-                cronProcessors.add(cronProcessor);
+                ScheduledFuture<?> scheduledFuture =
+                        scheduler.scheduleWithFixedDelay(fixedDelayProcessor, delay);
+                fixedDelayProcessors.add(fixedDelayProcessor);
                 scheduledFutures.add(scheduledFuture);
             } catch (Exception e) {
                 throw new DriverException(e);
@@ -104,8 +103,8 @@ public class CronDriverProvider implements DriverProvider {
                 for (ScheduledFuture<?> scheduledFuture : scheduledFutures) {
                     scheduledFuture.cancel(true);
                 }
-                for (CronProcessor cronProcessor : cronProcessors) {
-                    cronProcessor.shutdown();
+                for (FixedDelayProcessor fixedDelayProcessor : fixedDelayProcessors) {
+                    fixedDelayProcessor.shutdown();
                 }
             } finally {
                 lock.unlock();
@@ -113,9 +112,9 @@ public class CronDriverProvider implements DriverProvider {
         }
     }
 
-    private static class CronProcessor implements Runnable {
+    private static class FixedDelayProcessor implements Runnable {
 
-        private static final Logger LOGGER = LoggerFactory.getLogger(CronProcessor.class);
+        private static final Logger LOGGER = LoggerFactory.getLogger(FixedDelayProcessor.class);
 
         private final EvaluateService evaluateService;
         private final LongIdKey sectionKey;
@@ -123,7 +122,7 @@ public class CronDriverProvider implements DriverProvider {
         private final Lock lock = new ReentrantLock();
         private boolean runningFlag = true;
 
-        private CronProcessor(EvaluateService evaluateService, LongIdKey sectionKey) {
+        private FixedDelayProcessor(EvaluateService evaluateService, LongIdKey sectionKey) {
             this.evaluateService = evaluateService;
             this.sectionKey = sectionKey;
         }
@@ -136,7 +135,7 @@ public class CronDriverProvider implements DriverProvider {
                     return;
                 }
 
-                LOGGER.debug("计划时间已到达, cron驱动器驱动 " + sectionKey + " 部件执行评估动作...");
+                LOGGER.debug("计划时间已到达, fixed delay 驱动器驱动 " + sectionKey + " 部件执行评估动作...");
                 evaluateService.evaluate(sectionKey);
             } catch (Exception e) {
                 LOGGER.warn("记录 " + sectionKey + " 时出现异常, 放弃本次记录", e);